remove dependency on `glob` dependency
This resolves npm-audit warnings for minimatch:
* https://github.com/advisories/GHSA-hxm2-r34f-qmc5
* https://github.com/advisories/GHSA-f8q6-p94x-37v3

which indirectly affect glob 3.0.0 - 5.0.14.

* https://github.com/isaacs/node-glob/blob/v8.0.0/changelog.md

glob is quite large and pulls in a dozen dependencies, several of
which have dropped support for various Node versions, which would make
this patch a semver-major breaking change.

gilded-wordpress only uses this to non-recursively list the files
inside a single directory. This is easy to do with the Node API
nowadays.

Differences:
* glob() ignores inexistence, fs throws. Catch error.
* glob() returns absolute path, fs relative. Join accordingly.